Don't stop md because it returns CTP(77).
Test both simulation and real front, it all return CTP(77). And the next login will fail. So just don't stop md until it disconnects. Or it never disconnects.

If it disconnects, the internal state sill change, so it knows to start md. If it doesn't disconnect, it keeps login state. So nothing to do.