Fix / zero exception.
During reconnect, instruments' info is cleared and then re-query, and causes the qry task to access an empty container, leading to exception.

So we don't clear the instrument info in reconnect. Only clear them in login, along with setting last qry mark to false. So the instruments' info is always available, no exception.